requirements.txt
selenium==4.19.0
beautifulsoup4
requests
pyyaml